Fix test and split it to avoid timeouts
