+++ trunk/examples/programs/FloatingPoint/regression/c/noMathSAT/ctrans-float-remainder-zero.c
@@ -1,0 +1,20 @@
+//#Safe
+/*
+	https://en.cppreference.com/w/c/numeric/math/remainder
+*/
+
+#include <math.h>
+
+extern void __VERIFIER_error() __attribute__ ((__noreturn__));
+void __VERIFIER_assert(int cond) { if (!(cond)) { ERROR: __VERIFIER_error(); } return; }
+
+int main(void)
+{
+  double r = remainder(0.0, 1);
+	__VERIFIER_assert(r == 0 && !signbit(r));
+  r = remainder(-0.0, 1);
+	__VERIFIER_assert(r == 0 && signbit(r));
+
+	return 0;
+}
+
